refactor(fdl): Remove ktx imports
